refactor(archive): Simplify archiver test suite
Cleans up the archiver test suite that tested data store getters, so no
L1 data is required. Also extracts common test factory methods to a
separate file.
