+++ yarn-project/archiver/src/test/index.ts
@@ -1,3 +1,4 @@
+export * from './mock_structs.js';
 export * from './mock_l2_block_source.js';
 export * from './mock_l1_to_l2_message_source.js';
 export * from './mock_archiver.js';
